// gestimowebfront-end/src/gs-api/src/models/bail-villa-dto.ts
/* tslint:disable */
export interface BailVillaDto {
  id?: number;
  idAgence?: number;
  idCreateur?: number;
  designationBail?: string;
  abrvCodeBail?: string;
  enCoursBail?: boolean;
  archiveBail?: boolean;
  montantCautionBail?: number;
  nbreMoisCautionBail?: number;
  nouveauMontantLoyer?: number;
  dateDebut?: string;
  dateFin?: string;
  idVilla?: number;
  idLocataire?: number;
  fullNomLocatire?: string;
  codeBien?: string;
}