// gestimowebfront-end/src/gs-api/src/models/villa-dto.ts
/* tslint:disable */
export interface VillaDto {
  id?: number;
  idAgence?: number;
  idCreateur?: number;
  nbrePieceVilla?: number;
  nbrChambreVilla?: number;
  nbrSalonVilla?: number;
  nbrSalleEauVilla?: number;
  numVilla?: number;
  codeAbrvBienImmobilier?: string;
  nomCompletBienImmobilier?: string;
  nomBaptiserBienImmobilier?: string;
  description?: string;
  superficieBien?: number;
  bienMeublerResidence?: boolean;
  idSite?: number;
  idUtilisateur?: number;
  proprietaire?: string;
  idChapitre?: number;
  occupied?: boolean;
}